+++ wemake_python_styleguide/violations/best_practices.py
@@ -37,6 +37,7 @@ Summary
    BadMagicMethodViolation
    NestedImportViolation
    RedundantForElseViolation
+   ReassigningVariableToItselfViolation
    RedundantFinallyViolation
 
 Comments
@@ -69,6 +70,7 @@ Design
 .. autoclass:: BadMagicMethodViolation
 .. autoclass:: NestedImportViolation
 .. autoclass:: RedundantForElseViolation
+.. autoclass:: ReassigningVariableToItselfViolation
 .. autoclass:: RedundantFinallyViolation
 
 """
@@ -669,3 +671,19 @@ class RedundantFinallyViolation(ASTViolation):
     #: Error message shown to the user.
     error_template = 'Found `finally` in `try` block without `except`'
     code = 437
+
+
+class ReassigningVariableToItselfViolation(ASTViolation):
+    """
+    Forbids to assign variable to itself.
+
+    Reasoning:
+        There is no need to do that.
+
+    Note:
+        Returns Z438 as error code
+    """
+
+    #: Error message shown to the user.
+    error_template = 'Found reassigning variable "{0}" to itself'
+    code = 438